Metadata/requirements.txt
Pillow==9.2.0
pdfx==1.4.1